Make asserts from jsoncpp more descriptive
